+++ src/communicate/_utils/handlers/callback_handlers.py
@@ -134,7 +134,7 @@ def handle_search_follow_mode(update_params: UpdateBasicParams, extra_params: Up
         bot_message = 'Режим выбора поисков для отслеживания отключен. Переоткройте список активных поисков.'
 
     elif user_callback and user_callback.action == 'search_follow_clear':
-        db().delete_search_follow_marks(user_id)
+        db().delete_search_whiteness(user_id)
         bot_message = 'Все пометки отслеживания поисков сброшены. Переоткройте список активных поисков.'
 
     tg_api().send_callback_answer_to_api(update_params.user_id, callback_id, bot_message)
